FakeFinder_mybackend-main/requirements.txt
fastapi
uvicorn
pytesseract
pillow
playwright
transformers
sentence-transformers
torch
openai
requests